requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-silk==4.1.0
djangorestframework==3.12.2
gprof2dot==2021.2.21
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
wcwidth==0.2.5
zope.interface==5.4.0
